done, norminette now

+++ libft/ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: amoussai <amoussai@student.42.fr>          +#+  +:+       +#+        */
+/*   By: amoussai <amoussai@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 17:08:59 by amoussai          #+#    #+#             */
-/*   Updated: 2019/11/02 09:39:34 by amoussai         ###   ########.fr       */
+/*   Updated: 2021/03/06 12:25:12 by amoussai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,9 +25,9 @@ static char		*ft_isspace(char *str)
 	return (str);
 }
 
-int				ft_atoi(const char *str)
+long long				ft_atoi(const char *str)
 {
-	int	res;
+	long long	res;
 	int	neg;
 	int x;
